requirements.txt
discord.py[voice]==2.7.1
python-dotenv==1.1.0
lavalink==5.11.0
